[BE] fix: 게시판 쿼리 수정, 예외 처리 추가
