+++ server/src/main/java/com/codestates/server/board/controller/BoardController.java
@@ -3,11 +3,9 @@ package com.codestates.server.board.controller;
 import com.codestates.server.board.assembler.BoardAssembler;
 import com.codestates.server.board.dto.BoardDto;
 import com.codestates.server.board.entity.Board;
-import com.codestates.server.board.mapper.BoardMapper;
 import com.codestates.server.board.service.BoardService;
 import com.codestates.server.dto.MultiResponseDto;
-import com.codestates.server.exception.CustomException;
-import com.codestates.server.exception.ExceptionCode;
+import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
@@ -20,41 +18,30 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
+@AllArgsConstructor
 @RestController
 @RequestMapping("/board")
 @Validated
 public class BoardController {
 
     private final BoardService boardService;
-
-    private final BoardMapper mapper;
-
     private final BoardAssembler assembler;
-
-    public BoardController(BoardService boardService, BoardMapper mapper, BoardAssembler assembler) {
-        this.boardService = boardService;
-        this.mapper = mapper;
-        this.assembler = assembler;
-    }
 
     @GetMapping("/{id}")
     public EntityModel<BoardDto.Response> getBoard(@PathVariable long id) {
-        Board board = boardService.findOne(id);
-        BoardDto.Response response = mapper.boardToBoardResponseDto(board);
-        return assembler.toModel(response);
+        return assembler.toModel(boardService.findOne(id));
     }
 
     @GetMapping
     public ResponseEntity getBoardsPaged(@Positive @RequestParam int page,
-                                    @Positive @RequestParam int size) {
+                                         @Positive @RequestParam int size) {
 
         Page<Board> pagedBoards = boardService.findAllByPage(page - 1, size);
-        List<EntityModel<BoardDto.Response>> boards = boardStream(pagedBoards.getContent());
+        List<EntityModel<BoardDto.Response>> boards = boardService.boardStream(pagedBoards.getContent());
 
         return new ResponseEntity<>(
                 new MultiResponseDto<>(boards, pagedBoards), HttpStatus.OK);
@@ -65,7 +52,7 @@ public class BoardController {
     public CollectionModel<EntityModel<BoardDto.Response>> getBoards() {
 
         List<Board> boards = boardService.findAll();
-        List<EntityModel<BoardDto.Response>> response = boardStream(boards);
+        List<EntityModel<BoardDto.Response>> response = boardService.boardStream(boards);
 
         return CollectionModel.of(response,
                 linkTo(methodOn(BoardService.class).findAll()).withSelfRel());
@@ -75,6 +62,7 @@ public class BoardController {
     public ResponseEntity<?> postBoard(@Valid @RequestBody BoardDto.Post requestBody) {
 
         EntityModel<BoardDto.Response> entityModel = assembler.toModel(boardService.createOne(requestBody));
+
         return ResponseEntity
                 .created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri())
                 .body(entityModel);
@@ -95,13 +83,8 @@ public class BoardController {
     public ResponseEntity<?> likeBoard(@PathVariable("id") @Positive long id,
                                        @PathVariable("like") String like) {
 
-        Board board = boardService.findVerifiedBoard(id);
-
         char operator = like.equals("like") ? '1' : like.equals("dislike") ? '0' : 'x';
-        if (operator == 'x') throw new CustomException(ExceptionCode.BOARD_URL_NOT_FOUND);
-
-        boardService.changeLike(board, operator);
-        EntityModel<BoardDto.Response> entityModel = assembler.toModel(mapper.boardToBoardResponseDto(board));
+        EntityModel<BoardDto.Response> entityModel = assembler.toModel(boardService.changeLike(id, operator));
 
         return ResponseEntity
                 .created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri())
@@ -119,20 +102,12 @@ public class BoardController {
                                          @PathVariable("tag") String tag) {
 
         char operator = tag.equals("post") ? 'p' : tag.equals("asset") ? 'a' : 'x';
-        if (operator == 'x') throw new CustomException(ExceptionCode.BOARD_TAG_NOT_FOUND);
 
         Page<Board> pagedBoards = boardService.findAllByTag(page - 1, size, operator);
-        List<EntityModel<BoardDto.Response>> boards = boardStream(pagedBoards.getContent());
+        List<EntityModel<BoardDto.Response>> boards = boardService.boardStream(pagedBoards.getContent());
 
         return new ResponseEntity<>(
                 new MultiResponseDto<>(boards, pagedBoards), HttpStatus.OK);
-    }
-
-    public List<EntityModel<BoardDto.Response>> boardStream(List<Board> listedBoards) {
-        return listedBoards.stream()
-                .map(mapper::boardToBoardResponseDto)
-                .map(assembler::toModel)
-                .collect(Collectors.toList());
     }
 
 }
