+++ server/src/main/java/com/codestates/server/board/service/BoardService.java
@@ -1,18 +1,19 @@
 package com.codestates.server.board.service;
 
 
+import com.codestates.server.board.assembler.BoardAssembler;
 import com.codestates.server.board.dto.BoardDto;
 import com.codestates.server.board.entity.Board;
 import com.codestates.server.board.mapper.BoardMapper;
 import com.codestates.server.board.repository.BoardRepository;
 import com.codestates.server.comment.entity.Comment;
 import com.codestates.server.comment.repository.CommentRepository;
-import com.codestates.server.exception.BusinessLogicException;
 import com.codestates.server.exception.CustomException;
 import com.codestates.server.exception.ExceptionCode;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.hateoas.EntityModel;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,6 +21,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Transactional(readOnly = true)
@@ -28,15 +30,16 @@ public class BoardService {
 
     private final BoardRepository repository;
     private final BoardMapper mapper;
+    private final BoardAssembler assembler;
     private final CommentRepository commentRepository;
 
-    public Board findOne(long id) {
+    public BoardDto.Response findOne(long id) {
         Board verifiedBoard = findVerifiedBoard(id);
-        if (verifiedBoard.getBoardStatus() == Board.BoardStatus.BOARD_DELETED) {
-            throw new BusinessLogicException(ExceptionCode.BOARD_NOT_FOUND);
-        }
+
+        // deleted check
+
         repository.updateView(id);
-        return verifiedBoard;
+        return mapper.boardToBoardResponseDto(verifiedBoard);
     }
 
     public List<Board> findAll() {
@@ -48,8 +51,11 @@ public class BoardService {
     }
 
     public Page<Board> findAllByTag(int page, int size, char operator) {
+
+        if (operator == 'x') throw new CustomException(ExceptionCode.BOARD_TAG_NOT_FOUND);
+
         return operator == 'p' ? repository.findAllPost(PageRequest.of(page, size))
-                : repository.findAllAssetPost(PageRequest.of(page, size));
+                :  repository.findAllAssetPost(PageRequest.of(page, size));
     }
 
     @Transactional
@@ -84,16 +90,20 @@ public class BoardService {
     }
 
     @Transactional
-    public Board changeLike(Board board, char operator) {
-        Board verifiedBoard = findVerifiedBoard(board.getBoardId());
-        int like = board.getLike();
+    public BoardDto.Response changeLike(long id, char operator) {
+
+        if (operator == 'x') throw new CustomException(ExceptionCode.BOARD_URL_NOT_FOUND);
+
+        Board verifiedBoard = findVerifiedBoard(id);
+        int like = verifiedBoard.getLike();
 
         if (operator == '1') {
             verifiedBoard.setLike(++like);  // increase 1 like
         } else {
             verifiedBoard.setLike(like > 0 ? --like : 0);  // decrease 1 like (no - value)
         }
-        return repository.save(verifiedBoard);
+
+        return mapper.boardToBoardResponseDto(repository.save(verifiedBoard));
     }
 
     @Transactional
@@ -108,6 +118,13 @@ public class BoardService {
         commentRepository.deleteAllInBatch(relatedComments);
     }
 
+    public List<EntityModel<BoardDto.Response>> boardStream(List<Board> listedBoards) {
+        return listedBoards.stream()
+                .map(mapper::boardToBoardResponseDto)
+                .map(assembler::toModel)
+                .collect(Collectors.toList());
+    }
+
     public Board findVerifiedBoard(long id) {
         Optional<Board> optionalBoard = repository.findById(id);
         Board board = optionalBoard.orElseThrow(() -> new CustomException(ExceptionCode.BOARD_NOT_FOUND));
@@ -118,24 +135,19 @@ public class BoardService {
         return board;
     }
 
-//    public List<EntityModel<BoardDto.Response>> boardStream(List<Board> listedBoards) {
-//        return listedBoards.stream()
-//                .map(mapper::boardToBoardResponseDto)
-//                .map(assembler::toModel)
-//                .collect(Collectors.toList());
-//    }
-
     public Board.BoardTag verifyTag(Object board) {
 
         String oldTag = null;
         Board.BoardTag newTag = null;
 
+        // Check classes
         if (board instanceof BoardDto.Post && ((BoardDto.Post) board).getTag() != null) {
             oldTag = ((BoardDto.Post) board).getTag();
         } else if (board instanceof BoardDto.Patch) {
             oldTag = ((BoardDto.Patch) board).getTag();
         }
 
+        // Tag loop
         for (Board.BoardTag t : Board.BoardTag.values()) {
             if (t.getTag().equals(oldTag)) {
                 newTag = t;
